+++ rfc9535/test_suite/results.csv
@@ -1,5 +1,4 @@
 Total; Passed; Failed; Date
-687; 631; 48; 2025-03-10 21:36:58
 687; 631; 48; 2025-03-10 21:37:32
 687; 636; 43; 2025-03-10 21:56:27
 687; 636; 43; 2025-03-11 22:03:14
@@ -9,3 +8,4 @@ Total; Passed; Failed; Date
 687; 640; 37; 2025-03-11 22:37:58
 687; 642; 35; 2025-03-11 22:51:22
 687; 642; 35; 2025-03-11 22:54:09
+687; 642; 35; 2025-03-12 23:25:37
