+++ src/parser/macros.rs
@@ -1,121 +1,222 @@
-#[macro_export]
-macro_rules! filter {
-   () => {FilterExpression::Atom(op!,FilterSign::new(""),op!())};
-   ( $left:expr, $s:literal, $right:expr) => {
-      FilterExpression::Atom($left,FilterSign::new($s),$right)
-   };
-   ( $left:expr,||, $right:expr) => {FilterExpression::Or(Box::new($left),Box::new($right)) };
-   ( $left:expr,&&, $right:expr) => {FilterExpression::And(Box::new($left),Box::new($right)) };
-   ( count $inner:expr  ) => { FilterExpression::Extension(FilterExt::Count, vec![filter!(op!($inner),"exists",op!())])};
-   ( length $inner:expr  ) =>  { FilterExpression::Extension(FilterExt::Length, vec![filter!(op!($inner),"exists",op!())])};
-   ( value $inner:expr  ) => { FilterExpression::Extension(FilterExt::Value, vec![filter!(op!($inner),"exists",op!())])};
-   ( search  $inner1:expr,$inner2:expr  ) => { FilterExpression::Extension(FilterExt::Search,vec![
-       filter!(op!($inner1),"exists",op!()),
-       filter!(op!($inner2),"exists",op!())
-   ])};
-   ( match_  $inner1:expr,$inner2:expr  ) => { FilterExpression::Extension(FilterExt::Match,vec![
-       filter!(op!($inner1),"exists",op!()),
-       filter!(op!($inner2),"exists",op!())
-   ])};
-}
+use crate::parser::model::{Comparable, Filter, FilterAtom, FnArg, Literal, Segment, SingularQuery, Test};
 
 #[macro_export]
-macro_rules! op {
-    ( ) => {
-        Operand::Dynamic(Box::new(JsonPath::Empty))
+macro_rules! lit {
+    () => {
+        Literal::Null
     };
-    ( $s:literal) => {
-        Operand::Static(json!($s))
+    (b$b:expr ) => {
+        Literal::Bool($b)
     };
-    ( s $s:expr) => {
-        Operand::Static(json!($s))
+    (s$s:expr) => {
+        Literal::String($s.to_string())
     };
-    ( $s:expr) => {
-        Operand::Dynamic(Box::new($s))
+    (i$n:expr) => {
+        Literal::Int($n)
+    };
+    (f$n:expr) => {
+        Literal::Float($n)
     };
 }
 
 #[macro_export]
-macro_rules! idx {
-   ( $s:literal) => {JsonPathIndex::Single(json!($s))};
-   ( idx $($ss:literal),+) => {{
-       let ss_vec = vec![
-           $(
-               json!($ss),
-           )+
-       ];
-       JsonPathIndex::UnionIndex(ss_vec)
-   }};
-   ( $($ss:literal),+) => {{
-       let ss_vec = vec![
-           $(
-               $ss.to_string(),
-           )+
-       ];
-       JsonPathIndex::UnionKeys(ss_vec)
-   }};
-   ( $s:literal) => {JsonPathIndex::Single(json!($s))};
-   ( ? $s:expr) => {JsonPathIndex::Filter($s)};
-   ( [$l:literal;$m:literal;$r:literal]) => {JsonPathIndex::Slice(Some($l),Some($m),Some($r))};
-   ( [$l:literal;$m:literal;]) => {JsonPathIndex::Slice($l,$m,1)};
-   ( [$l:literal;;$m:literal]) => {JsonPathIndex::Slice($l,0,$m)};
-   ( [;$l:literal;$m:literal]) => {JsonPathIndex::Slice(0,$l,$m)};
-   ( [;;$m:literal]) => {JsonPathIndex::Slice(None,None,Some($m))};
-   ( [;$m:literal;]) => {JsonPathIndex::Slice(None,Some($m),None)};
-   ( [$m:literal;;]) => {JsonPathIndex::Slice(Some($m),None,None)};
-   ( [;;]) => {JsonPathIndex::Slice(None,None,None)};
+macro_rules! q_segments {
+    ($segment:tt) => {
+        vec![q_segment!($segment)]
+    };
+    // Recursive case: multiple segments
+    ($segment:tt $($rest:tt)*) => {{
+        let mut segments = q_segments!($($rest)*);
+        segments.insert(0, q_segment!($segment));
+        segments
+    }};
 }
 
 #[macro_export]
-macro_rules! chain {
-    ($($ss:expr),+) => {{
-        let ss_vec = vec![
-            $(
-                $ss,
-            )+
-        ];
-        JsonPath::Chain(ss_vec)
-   }};
+macro_rules! q_segment {
+    ($name:ident) => {
+        SingularQuerySegment::Name(stringify!($name).to_string())
+    };
+    ([$name:ident]) => {
+        SingularQuerySegment::Name(format!("\"{}\"", stringify!($name)))
+    };
+    ([$index:expr]) => {
+        SingularQuerySegment::Index($index)
+    };
 }
-
-/// Can be used to Parse a JsonPath with a more native way.
-/// e.g.
-/// ```rust
-/// use jsonpath_rust::{path, JsonPath};
-/// use std::str::FromStr;
-/// use serde_json::Value;
-///
-/// let path:JsonPath<Value> = JsonPath::from_str(".abc.*").unwrap();
-/// let path2 = JsonPath::Chain(vec![path!("abc"), path!(*)]);
-/// assert_eq!(path, path2);
-/// ```
 #[macro_export]
-macro_rules! path {
-   ( ) => {JsonPath::Empty};
-   (*) => {JsonPath::Wildcard};
-   ($) => {JsonPath::Root};
-   (@) => {JsonPath::Current(Box::new(JsonPath::Empty))};
-   (@$e:expr) => {JsonPath::Current(Box::new($e))};
-   (@,$($ss:expr),+) => {{
-       let ss_vec = vec![
-           $(
-               $ss,
-           )+
-       ];
-       let chain = JsonPath::Chain(ss_vec);
-       JsonPath::Current(Box::new(chain))
-   }};
-   (..$e:literal) => {JsonPath::Descent($e.to_string())};
-   (..*) => {JsonPath::DescentW};
-   ($e:literal) => {JsonPath::Field($e.to_string())};
-   ($e:expr) => {JsonPath::Index($e)};
+macro_rules! singular_query {
+    (@ $($segment:tt)*) => {
+        SingularQuery::Current(q_segments!($($segment)*))
+    };
+    ($($segment:tt)*) => {
+        SingularQuery::Root(q_segments!($($segment)*))
+    };
 }
 
-#[cfg(test)]
-pub(crate) use chain;
-#[cfg(test)]
-pub(crate) use filter;
-#[cfg(test)]
-pub(crate) use idx;
-#[cfg(test)]
-pub(crate) use op;
+#[macro_export]
+macro_rules! slice {
+    () => {
+        (None, None, None)
+    };
+    ($start:expr) => {
+        (Some($start), None, None)
+    };
+    ($start:expr, $end:expr) => {
+        (Some($start), Some($end), None)
+    };
+    ($start:expr,, $step:expr) => {
+        (Some($start), None, Some($step))
+    };
+    (,, $step:expr) => {
+        (None, None, Some($step))
+    };
+    (, $end:expr) => {
+        (None, Some($end), None)
+    };
+    (, $end:expr,$step:expr ) => {
+        (None, Some($end), Some($step))
+    };
+    ($start:expr, $end:expr, $step:expr) => {
+        (Some($start), Some($end), Some($step))
+    };
+}
+
+#[macro_export]
+macro_rules! test_fn {
+    ($name:ident $arg:expr) => {
+        TestFunction::try_new(stringify!($name), vec![$arg]).unwrap()
+    };
+
+    ($name:ident $arg1:expr, $arg2:expr ) => {
+        TestFunction::try_new(stringify!($name), vec![$arg1, $arg2]).unwrap()
+    };
+}
+
+#[macro_export]
+macro_rules! arg {
+    ($arg:expr) => {
+        FnArg::Literal($arg)
+    };
+    (t $arg:expr) => {
+        FnArg::Test(Box::new($arg))
+    };
+    (f $arg:expr) => {
+        FnArg::Filter($arg)
+    }
+}
+
+#[macro_export]
+macro_rules! test {
+  (@ $($segments:expr)*) => { Test::RelQuery(vec![$($segments),*]) };
+  (S $jq:expr) => { Test::AbsQuery($jq) };
+  ($tf:expr) => { Test::Function(Box::new($tf)) };
+
+}
+
+#[macro_export]
+macro_rules! or {
+    ($($items:expr),*) => {
+       crate::parser::model::Filter::Or(vec![ $($items),* ])
+    };
+}
+
+#[macro_export]
+macro_rules! and {
+    ($($items:expr),*) => {
+        crate::parser::model::Filter::And(vec![ $($items),* ])
+    };
+}
+
+#[macro_export]
+macro_rules! filter_ {
+    ($item:expr) => {
+        crate::parser::model::Filter::Atom($item)
+    };
+
+    (or $($items:expr),*) => {
+       crate::parser::model::Filter::Or(vec![ $($items),* ])
+    };
+
+    (and $($items:expr),*) => {
+        crate::parser::model::Filter::And(vec![ $($items),* ])
+    };
+}
+
+#[macro_export]
+macro_rules! atom {
+    (! $filter:expr) => {
+        FilterAtom::filter($filter, true)
+    };
+    ($filter:expr) => {
+        FilterAtom::filter($filter, false)
+    };
+    (t! $filter:expr) => {
+        FilterAtom::test($filter, true)
+    };
+    (t $filter:expr) => {
+        FilterAtom::filter($filter, false)
+    };
+    ($lhs:expr, $s:expr, $rhs:expr) => {
+        FilterAtom::Comparison(Box::new(cmp!($lhs, $s, $rhs)))
+    };
+}
+
+#[macro_export]
+macro_rules! cmp {
+  ($lhs:expr, $op:expr , $rhs:expr) => {
+      Comparison::try_new($op, $lhs, $rhs).unwrap()
+  }
+}
+
+#[macro_export]
+macro_rules! comparable {
+    ($lit:expr) => {
+        Comparable::Literal($lit)
+    };
+    (f $func:expr) => {
+        Comparable::Function($func)
+    };
+    (> $sq:expr) => {
+        Comparable::SingularQuery($sq)
+    };
+}
+
+#[macro_export]
+macro_rules! selector {
+    (*) => {
+        Selector::Wildcard
+    };
+    (?$filter:expr) => {
+        Selector::Filter($filter)
+    };
+    ($name:ident) => {
+        Selector::Name(stringify!($name).to_string())
+    };
+    ([$name:ident]) => {
+        Selector::Name(format!("\"{}\"", stringify!($name)))
+    };
+    ([$index:expr]) => {
+        Selector::Index($index)
+    };
+}
+
+#[macro_export]
+macro_rules! segment {
+    (..$segment:expr) => {
+        Segment::Descendant(Box::new($segment))
+    };
+    ($selector:expr) => {
+        Segment::Selector($selector)
+    };
+    ($($selectors:expr),*) => {
+        Segment::Selectors(vec![$($selectors),*])
+    };
+}
+
+#[macro_export]
+macro_rules! jq {
+    ($($segment:expr),*) => {
+        JpQuery::new(vec![$($segment),*])
+    };
+}
